# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "SocialAnimal REST", "version": "1.0.0" },
	"components": {
		"securitySchemes": {
			"bearerAuth": { "type": "http", "scheme": "bearer" }
		}
	},
	"paths": {
...
		"/api/v1/calendars": {
			"get": {
				"operationId": "calendars",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 500,
							"default": 100
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"type": "object",
												"additionalProperties": false,
												"properties": {
													"id": { "type": "string" },
													"name": {
														"type": "string"
													},
													"type": {
														"type": "string"
													},
													"connectionId": {
														"type": [
															"null",
															"string"
														]
													},
													"remoteId": {
														"type": [
															"null",
															"string"
														]
													},
													"syncInterval": {
														"type": "integer"
													},
													"eventCount": {
														"type": "integer"
													},
													"lastSync": {
														"type": [
															"null",
															"string"
														]
													},
													"lastAttempt": {
														"type": [
															"null",
															"string"
														]
													},
													"lastSuccess": {
														"type": [
															"null",
															"string"
														]
													},
													"lastError": {
														"type": [
															"null",
															"string"
														]
													}
												},
												"required": [
													"id",
													"name",
													"type",
													"syncInterval"
												]
											}
										},
										"nextCursor": {
											"type": ["null", "string"]
										}
									},
									"required": ["items", "nextCursor"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/calendars/{id}/sync-runs": {
			"post": {
				"operationId": "submitSync",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"202": {
						"description": "202",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"id": { "type": "string" },
										"calendarId": { "type": "string" },
										"status": { "type": "string" },
										"eventsSynced": {
											"type": ["null", "integer"]
										},
										"error": { "type": ["null", "string"] },
										"statusUrl": { "type": "string" }
									},
									"required": ["id", "calendarId", "status"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/providers": {
			"get": {
				"operationId": "providers",
				"parameters": [],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"type": "object",
												"additionalProperties": false,
												"properties": {
													"id": { "type": "string" },
													"name": {
														"type": "string"
													},
													"discovery": {
														"type": "boolean"
													},
													"sync": {
														"type": "boolean"
													},
													"test": {
														"type": "boolean"
													},
													"oauth": {
														"type": "boolean"
													},
													"import": {
														"type": "boolean"
													}
												},
												"required": [
													"id",
													"name",
													"discovery",
													"sync",
													"test",
													"oauth",
													"import"
												]
											}
										},
										"nextCursor": {
											"type": ["null", "string"]
										}
									},
									"required": ["items", "nextCursor"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/connections/{id}": {
			"patch": {
				"operationId": "updateConnection",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"properties": {
									"name": { "type": "string" },
									"credentials": {
										"type": "object",
										"additionalProperties": false,
										"properties": {
											"url": { "type": "string" },
											"username": { "type": "string" },
											"password": { "type": "string" },
											"accessToken": { "type": "string" },
											"refreshToken": { "type": "string" }
										},
										"required": []
									}
								},
								"required": []
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"id": { "type": "string" },
										"name": { "type": "string" },
										"type": { "type": "string" },
										"hasCredentials": { "type": "boolean" }
									},
									"required": [
										"id",
										"name",
										"type",
										"hasCredentials"
									]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/connections/google/callback": {
			"get": {
				"operationId": "googleCallback",
				"parameters": [
					{
						"name": "state",
						"in": "query",
						"required": true,
						"schema": { "type": "string" }
					},
					{
						"name": "code",
						"in": "query",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"302": {
						"description": "302",
						"content": {
							"application/json": { "schema": { "type": "null" } }
						}
					}
				}
			}
		},
...
	}
}